// src/app/modules/admin/admin.interface.ts
// import { Model, Types } from 'mongoose';
// import { IManagementDepartment } from '../managementDepartment/managementDepartment.interface';

// export type UserName = {
//   firstName: string;
//   lastName: string;
//   middleName: string;
// };

// export type IAdmin = {
//   id: string;
//   name: UserName; //embedded object
//   gender: 'male' | 'female';
//   dateOfBirth: string;
//   email: string;
//   contactNo: string;
//   emergencyContactNo: string;
//   bloodGroup?: 'A+' | 'A-' | 'B+' | 'B-' | 'AB+' | 'AB-' | 'O+' | 'O-';
//   presentAddress: string;
//   permanentAddress: string;
//   profileImage?: string;
//   managementDepartment: Types.ObjectId | IManagementDepartment;
//   designation: string;
// };

// export type AdminModel = Model<IAdmin, Record<string, unknown>>;

// export type IAdminFilters = {
//   searchTerm?: string;
//   id?: string;
//   bloodGroup?: string;
//   email?: string;
//   contactNo?: string;
//   designation: string;
// };

import { Model, Types } from 'mongoose';
import { IManagementDepartment } from '../managementDepartment/managementDepartment.interface';

export type UserName = {
  firstName: string;
  lastName: string;
  middleName: string;
};

export type IAdmin = {
  id: string;
  name: UserName;
  profileImage: string;
  dateOfBirth?: string;
  email: string;
  contactNo: string;
  emergencyContactNo: string;
  gender?: 'male' | 'female';
  permanentAddress?: string;
  presentAddress?: string;
  bloodGroup?: 'A+' | 'A-' | 'B+' | 'B-' | 'AB+' | 'AB-' | 'O+' | 'O-';

  managementDepartment: Types.ObjectId | IManagementDepartment;
  designation: string;
};

export type AdminModel = Model<IAdmin, Record<string, unknown>>;

export type IAdminFilters = {
  searchTerm?: string;
  id?: string;
  email?: string;
  contactNo?: string;
  emergencyContactNo?: string;
  gender?: 'male' | 'female';
  bloodGroup?: 'A+' | 'A-' | 'B+' | 'B-' | 'AB+' | 'AB-' | 'O+' | 'O-';
  managementDepartment?: string;
  designation?: string;
};